Move LLVM intrinsics to their own module
